refactor(preparer): Move preparers from launcher-addon to core-impl
In order to include preparers dependencies, added those to core-impl:
- maven3-model-jdom-support
- commons-lang3

Remove Forge usage in `ChangeMavenMetadataPreparer`.

Extract ReadmeProcessor interface in api to make it accessible from the outside in `ProjectInfoStep` for example

new class ReadmePreparer to do the preparation logic and call to ReadmeProcessor when needed.

Move `add-template-script.sh` to core-impl and add it as part of the maven build.

Fixes #237

Co-authored-by: George Gastaldi <gegastaldi@gmail.com>