feat(osio): add Create/Delete space methods in OsioWitClient
Add methods createSpace and deleteSpace in OsioWitClient and use it in OsioWitClientTest.

Why:
It was not possible to clean codebases created by the tests since it's not possible to delete a codebase without its id.

Those new methods are visible even if they are only used by tests but it's ok since the osio-addon is not meant to be used as an api.