test(osio): Create/Delete test space in OsioEndpointIT
Why:
I was not possible to clean codebases created by the test since it's not posssible to delete a codebase without its id.

As a result:
It is cleaner since the space doesn't have to be created as a prerequisite and it's name doesn't need to be specified as env anymore.

Closes #287