fix: fixed tests to match the JAX-RS spec
Errors in the ParamConverters always return 404. See specification section "3.2 Fields and Bean Properties":

A WebApplicationException thrown during construction of field or property values using any of the 5 steps listed above is processed directly as described in Section 3.3.4.
Other exceptions thrown during construction of field or property values using any of the 5 steps listed above are treated as client errors:
if the field or property is annotated with @MatrixParam , @QueryParam or @PathParam then an implementation MUST generate an instance of NotFoundException (404 status) that
wraps the thrown exception and no entity; if the field or property is annotated with @HeaderParam or @CookieParam then an implementation
MUST generate an instance of BadRequestException (400 status) that wraps the thrown exception and no entity.

Reference: https://issues.jboss.org/browse/RESTEASY-1364