feat(analysis): The repo analysis now returns much more info
Besides what it returned before it now returns a list of all folders
in the repo that contain builable code and the builder image that would
be needed for each of them. A full list of folders is also included
in case the user wants to override the default folder with one of
their choosing.